Reorganize setup document for R development
This commit reorganizes the setup document, makes it specific for R
development, and also makes it easier for new users.

The package management installation process is put after the R package
download path and the introduction to the interactive session, as these
two steps are crucial in the package management steps. Additionally,
commentary is added for R specific installation during the package
management section.