Add input and quality control functions
These functions were used in a previous workflow analysis. Here I
formalize the inputs and add necessary documentation for future
reference and use.

This file is named for functions that will be used all up until the
start of the DADA2-specific functions in order to consolidate function
semantic types.