Create first input step in pipeline
I removed the `main()` function, which doesn't make sense in hindsight
as this pipeline requires human input. Thus, having a single function to
run the entire pipeline would never been used.

This commit adds the potential other steps of the pipeline that will be
filled in later on in the `dada2_pipeline.R` file. Documentation and
namespace definitions are created for this first step so far.