Insert R script within bash script to consolidate
Most often, we will probably need to write R scripts to be run during a
job submission. To make sure we don't have to create a submit script AND
a separate R script, here shows a minimal example of putting an R script
within a SLURM submit script.

The R scripts needed to use this package will generally be small so this
shouldn't create an unmanageable submit script.