Create variables for manually named files
In the test script for the filter, trimming, and error learning step,
lots of the file names and directories were written throughout the
script. However, to make these scripts easier to manage, I've put all
the variables and things that could be changed near the top in hopes to
make modifying the script easier for the user.